Create manual partial import process
As a temporary measure implement a manual process to allow the updating of the TDR realm configuration via the partial import functionality in the Keycloak GUI without wiping existing users. Complete solution will be to use the Keycloak REST APIs to do a partial import, and set up a Jenkins job to run the update.

Python script used to update the realm configuration file and save locally, so the realm configuration json can be imported into the Keycloak instance.

README updated with instructions for this process