Add a user monitoring task.
There is an existing scheduler within Keycloak which you can use. On the
postInit method, we wait for the post migration event. I'm not
completely sure why this event but it works.

We then create a new instance of the UserMonitoringTask which extends
ScheduledTask and carries out the checks against the users for all
realms.

If it finds users with no MFA, it sends a message to the notifications
SNS topic.