Use IAM authentication for the database.
I've added in a new class `RDSCredentialsProvider`

This is not a Keycloak SPI, it's an extension of the Quarkus server that
Keycloak runs on so it doesn't follow the usual SPI structure. This is
why it needs its own quarkus.properties file

https://quarkus.io/guides/credentials-provider#custom-credentials-provider

The rest is adding the jar to the Docker image and some new
dependencies.