Filter service account users from the MFA check.
Keycloak creates users for the service accounts and these are used by
things like the export and the backend checks and so can't be removed.

You can't log in as these users so they're safe without MFA so this
excludes them. There is a getServiceAccountClientLink method that is
null for non service accounts so we're filtering on that.

I've also replace the getUsersStream method with searchForUserStream as
getUsersStream is deprecated. I've updated the tests to reflect this and
added a test for the service account filter.