Change the element link for the error.
We were linking the error at the top of the page to #error-kc-form-login
which is the error on each individual page. The accessibility audit
found that we need to focus on the first input in the form when the
error link is clicked.

There's no way I can see to find out inside the templates which element
is the first in the form with the error and the error itself is in the
template so I've added in another parameter for the error target and
passed this in. It defaults to the value it was before so we'll have to
remember to pass it in if we ever do any more auth pages.

I've also moved the error above the h1 and added Error: to the page
title as per the report.